+++ Toggl.Foundation/DataSources/TimeEntriesDataSource.cs
@@ -37,7 +37,7 @@ namespace Toggl.Foundation.DataSources
         public IObservable<IEnumerable<IDatabaseTimeEntry>> GetAll()
             => repository.GetAll(te => !te.IsDeleted);
 
-        public IObservable<IDatabaseTimeEntry> GetById(int id)
+        public IObservable<IDatabaseTimeEntry> GetById(long id)
             => repository.GetById(id);
 
         public IObservable<Unit> Delete(long id)
